@@ -1,6 +1,6 @@
 module github.com/tradephantom/axcp-spec
 
-go 1.22.7
+go 1.23.4
 
 require (
 	github.com/quic-go/quic-go v0.49.0
