@@ -2,7 +2,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.0
+// 	protoc-gen-go v1.36.6
 // 	protoc        v6.31.0
 // source: axcp.proto
 
@@ -13,6 +13,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -255,14 +256,11 @@ func (DpParams_Granularity) EnumDescriptor() ([]byte, []int) {
 }
 
 type AxcpEnvelope struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Version uint32 `protobuf:"varint,1,opt,name=version,proto3" json:"version,omitempty"`               // protocol version (currently 1)
-	TraceId string `protobuf:"bytes,2,opt,name=trace_id,json=traceId,proto3" json:"trace_id,omitempty"` // request / conversation correlation
-	Profile uint32 `protobuf:"varint,3,opt,name=profile,proto3" json:"profile,omitempty"`               // 0 = Basic … 3 = Enterprise-Privacy
-	// Types that are assignable to Payload:
+	state   protoimpl.MessageState `protogen:"open.v1"`
+	Version uint32                 `protobuf:"varint,1,opt,name=version,proto3" json:"version,omitempty"`               // protocol version (currently 1)
+	TraceId string                 `protobuf:"bytes,2,opt,name=trace_id,json=traceId,proto3" json:"trace_id,omitempty"` // request / conversation correlation
+	Profile uint32                 `protobuf:"varint,3,opt,name=profile,proto3" json:"profile,omitempty"`               // 0 = Basic … 3 = Enterprise-Privacy
+	// Types that are valid to be assigned to Payload:
 	//
 	//	*AxcpEnvelope_ContextPatch
 	//	*AxcpEnvelope_CapabilityMsg
@@ -275,15 +273,15 @@ type AxcpEnvelope struct {
 	Payload          isAxcpEnvelope_Payload `protobuf_oneof:"payload"`
 	Signature        []byte                 `protobuf:"bytes,100,opt,name=signature,proto3" json:"signature,omitempty"`                                       // detached sig (profile ≥1)
 	AttestationProof []byte                 `protobuf:"bytes,101,opt,name=attestation_proof,json=attestationProof,proto3" json:"attestation_proof,omitempty"` // SGX / SEV quote (profile ≥2)
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *AxcpEnvelope) Reset() {
 	*x = AxcpEnvelope{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AxcpEnvelope) String() string {
@@ -294,7 +292,7 @@ func (*AxcpEnvelope) ProtoMessage() {}
 
 func (x *AxcpEnvelope) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -330,65 +328,81 @@ func (x *AxcpEnvelope) GetProfile() uint32 {
 	return 0
 }
 
-func (m *AxcpEnvelope) GetPayload() isAxcpEnvelope_Payload {
-	if m != nil {
-		return m.Payload
+func (x *AxcpEnvelope) GetPayload() isAxcpEnvelope_Payload {
+	if x != nil {
+		return x.Payload
 	}
 	return nil
 }
 
 func (x *AxcpEnvelope) GetContextPatch() *ContextPatch {
-	if x, ok := x.GetPayload().(*AxcpEnvelope_ContextPatch); ok {
-		return x.ContextPatch
+	if x != nil {
+		if x, ok := x.Payload.(*AxcpEnvelope_ContextPatch); ok {
+			return x.ContextPatch
+		}
 	}
 	return nil
 }
 
 func (x *AxcpEnvelope) GetCapabilityMsg() *CapabilityMessage {
-	if x, ok := x.GetPayload().(*AxcpEnvelope_CapabilityMsg); ok {
-		return x.CapabilityMsg
+	if x != nil {
+		if x, ok := x.Payload.(*AxcpEnvelope_CapabilityMsg); ok {
+			return x.CapabilityMsg
+		}
 	}
 	return nil
 }
 
 func (x *AxcpEnvelope) GetRouteMsg() *RoutePolicyMessage {
-	if x, ok := x.GetPayload().(*AxcpEnvelope_RouteMsg); ok {
-		return x.RouteMsg
+	if x != nil {
+		if x, ok := x.Payload.(*AxcpEnvelope_RouteMsg); ok {
+			return x.RouteMsg
+		}
 	}
 	return nil
 }
 
 func (x *AxcpEnvelope) GetError() *ErrorMessage {
-	if x, ok := x.GetPayload().(*AxcpEnvelope_Error); ok {
-		return x.Error
+	if x != nil {
+		if x, ok := x.Payload.(*AxcpEnvelope_Error); ok {
+			return x.Error
+		}
 	}
 	return nil
 }
 
 func (x *AxcpEnvelope) GetProfileNeg() *ProfileNegotiate {
-	if x, ok := x.GetPayload().(*AxcpEnvelope_ProfileNeg); ok {
-		return x.ProfileNeg
+	if x != nil {
+		if x, ok := x.Payload.(*AxcpEnvelope_ProfileNeg); ok {
+			return x.ProfileNeg
+		}
 	}
 	return nil
 }
 
 func (x *AxcpEnvelope) GetProfileAck() *ProfileAck {
-	if x, ok := x.GetPayload().(*AxcpEnvelope_ProfileAck); ok {
-		return x.ProfileAck
+	if x != nil {
+		if x, ok := x.Payload.(*AxcpEnvelope_ProfileAck); ok {
+			return x.ProfileAck
+		}
 	}
 	return nil
 }
 
 func (x *AxcpEnvelope) GetRetryEnv() *RetryEnvelope {
-	if x, ok := x.GetPayload().(*AxcpEnvelope_RetryEnv); ok {
-		return x.RetryEnv
+	if x != nil {
+		if x, ok := x.Payload.(*AxcpEnvelope_RetryEnv); ok {
+			return x.RetryEnv
+		}
 	}
 	return nil
 }
 
 func (x *AxcpEnvelope) GetTelemetry() *TelemetryDatagram {
-	if x, ok := x.GetPayload().(*AxcpEnvelope_Telemetry); ok {
-		return x.Telemetry
+	if x != nil {
+		if x, ok := x.Payload.(*AxcpEnvelope_Telemetry); ok {
+			return x.Telemetry
+		}
 	}
 	return nil
 }
@@ -460,23 +474,20 @@ func (*AxcpEnvelope_RetryEnv) isAxcpEnvelope_Payload() {}
 func (*AxcpEnvelope_Telemetry) isAxcpEnvelope_Payload() {}
 
 type DeltaOp struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Op            DeltaOp_OpType         `protobuf:"varint,1,opt,name=op,proto3,enum=axcp.v0_1.DeltaOp_OpType" json:"op,omitempty"`
+	Path          string                 `protobuf:"bytes,2,opt,name=path,proto3" json:"path,omitempty"` // JSON Pointer
+	Data          []byte                 `protobuf:"bytes,3,opt,name=data,proto3" json:"data,omitempty"` // gz-compressed payload
+	Ts            uint64                 `protobuf:"varint,4,opt,name=ts,proto3" json:"ts,omitempty"`    // lamport / microseconds
 	unknownFields protoimpl.UnknownFields
-
-	Op   DeltaOp_OpType `protobuf:"varint,1,opt,name=op,proto3,enum=axcp.v0_1.DeltaOp_OpType" json:"op,omitempty"`
-	Path string         `protobuf:"bytes,2,opt,name=path,proto3" json:"path,omitempty"` // JSON Pointer
-	Data []byte         `protobuf:"bytes,3,opt,name=data,proto3" json:"data,omitempty"` // gz-compressed payload
-	Ts   uint64         `protobuf:"varint,4,opt,name=ts,proto3" json:"ts,omitempty"`    // lamport / microseconds
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *DeltaOp) Reset() {
 	*x = DeltaOp{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *DeltaOp) String() string {
@@ -487,7 +498,7 @@ func (*DeltaOp) ProtoMessage() {}
 
 func (x *DeltaOp) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -531,22 +542,19 @@ func (x *DeltaOp) GetTs() uint64 {
 }
 
 type ContextPatch struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	ContextId     string                 `protobuf:"bytes,1,opt,name=context_id,json=contextId,proto3" json:"context_id,omitempty"`
+	BaseVersion   uint64                 `protobuf:"varint,2,opt,name=base_version,json=baseVersion,proto3" json:"base_version,omitempty"`
+	Ops           []*DeltaOp             `protobuf:"bytes,3,rep,name=ops,proto3" json:"ops,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	ContextId   string     `protobuf:"bytes,1,opt,name=context_id,json=contextId,proto3" json:"context_id,omitempty"`
-	BaseVersion uint64     `protobuf:"varint,2,opt,name=base_version,json=baseVersion,proto3" json:"base_version,omitempty"`
-	Ops         []*DeltaOp `protobuf:"bytes,3,rep,name=ops,proto3" json:"ops,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ContextPatch) Reset() {
 	*x = ContextPatch{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ContextPatch) String() string {
@@ -557,7 +565,7 @@ func (*ContextPatch) ProtoMessage() {}
 
 func (x *ContextPatch) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -594,21 +602,18 @@ func (x *ContextPatch) GetOps() []*DeltaOp {
 }
 
 type ContextGraphVersion struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	ContextId     string                 `protobuf:"bytes,1,opt,name=context_id,json=contextId,proto3" json:"context_id,omitempty"`
+	Version       uint64                 `protobuf:"varint,2,opt,name=version,proto3" json:"version,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	ContextId string `protobuf:"bytes,1,opt,name=context_id,json=contextId,proto3" json:"context_id,omitempty"`
-	Version   uint64 `protobuf:"varint,2,opt,name=version,proto3" json:"version,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ContextGraphVersion) Reset() {
 	*x = ContextGraphVersion{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ContextGraphVersion) String() string {
@@ -619,7 +624,7 @@ func (*ContextGraphVersion) ProtoMessage() {}
 
 func (x *ContextGraphVersion) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -649,20 +654,17 @@ func (x *ContextGraphVersion) GetVersion() uint64 {
 }
 
 type SyncSubscribe struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	From          *ContextGraphVersion   `protobuf:"bytes,1,opt,name=from,proto3" json:"from,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	From *ContextGraphVersion `protobuf:"bytes,1,opt,name=from,proto3" json:"from,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *SyncSubscribe) Reset() {
 	*x = SyncSubscribe{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SyncSubscribe) String() string {
@@ -673,7 +675,7 @@ func (*SyncSubscribe) ProtoMessage() {}
 
 func (x *SyncSubscribe) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -696,21 +698,18 @@ func (x *SyncSubscribe) GetFrom() *ContextGraphVersion {
 }
 
 type SyncRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	MissingFrom   *ContextGraphVersion   `protobuf:"bytes,1,opt,name=missing_from,json=missingFrom,proto3" json:"missing_from,omitempty"`
+	ToVersion     uint64                 `protobuf:"varint,2,opt,name=to_version,json=toVersion,proto3" json:"to_version,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	MissingFrom *ContextGraphVersion `protobuf:"bytes,1,opt,name=missing_from,json=missingFrom,proto3" json:"missing_from,omitempty"`
-	ToVersion   uint64               `protobuf:"varint,2,opt,name=to_version,json=toVersion,proto3" json:"to_version,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *SyncRequest) Reset() {
 	*x = SyncRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SyncRequest) String() string {
@@ -721,7 +720,7 @@ func (*SyncRequest) ProtoMessage() {}
 
 func (x *SyncRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -751,21 +750,18 @@ func (x *SyncRequest) GetToVersion() uint64 {
 }
 
 type RetryEnvelope struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	BufferedPatches []*ContextPatch `protobuf:"bytes,1,rep,name=buffered_patches,json=bufferedPatches,proto3" json:"buffered_patches,omitempty"`
-	TtlMs           uint32          `protobuf:"varint,2,opt,name=ttl_ms,json=ttlMs,proto3" json:"ttl_ms,omitempty"`
+	state           protoimpl.MessageState `protogen:"open.v1"`
+	BufferedPatches []*ContextPatch        `protobuf:"bytes,1,rep,name=buffered_patches,json=bufferedPatches,proto3" json:"buffered_patches,omitempty"`
+	TtlMs           uint32                 `protobuf:"varint,2,opt,name=ttl_ms,json=ttlMs,proto3" json:"ttl_ms,omitempty"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *RetryEnvelope) Reset() {
 	*x = RetryEnvelope{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RetryEnvelope) String() string {
@@ -776,7 +772,7 @@ func (*RetryEnvelope) ProtoMessage() {}
 
 func (x *RetryEnvelope) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -806,20 +802,17 @@ func (x *RetryEnvelope) GetTtlMs() uint32 {
 }
 
 type CapabilityOffer struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Desc          *CapabilityDescriptor  `protobuf:"bytes,1,opt,name=desc,proto3" json:"desc,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Desc *CapabilityDescriptor `protobuf:"bytes,1,opt,name=desc,proto3" json:"desc,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CapabilityOffer) Reset() {
 	*x = CapabilityOffer{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CapabilityOffer) String() string {
@@ -830,7 +823,7 @@ func (*CapabilityOffer) ProtoMessage() {}
 
 func (x *CapabilityOffer) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -853,20 +846,17 @@ func (x *CapabilityOffer) GetDesc() *CapabilityDescriptor {
 }
 
 type CapabilityRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Ids           []string               `protobuf:"bytes,1,rep,name=ids,proto3" json:"ids,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Ids []string `protobuf:"bytes,1,rep,name=ids,proto3" json:"ids,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CapabilityRequest) Reset() {
 	*x = CapabilityRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[8]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CapabilityRequest) String() string {
@@ -877,7 +867,7 @@ func (*CapabilityRequest) ProtoMessage() {}
 
 func (x *CapabilityRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[8]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -900,20 +890,17 @@ func (x *CapabilityRequest) GetIds() []string {
 }
 
 type CapabilityAck struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Accepted      []string               `protobuf:"bytes,1,rep,name=accepted,proto3" json:"accepted,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Accepted []string `protobuf:"bytes,1,rep,name=accepted,proto3" json:"accepted,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CapabilityAck) Reset() {
 	*x = CapabilityAck{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[9]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CapabilityAck) String() string {
@@ -924,7 +911,7 @@ func (*CapabilityAck) ProtoMessage() {}
 
 func (x *CapabilityAck) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[9]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -947,25 +934,22 @@ func (x *CapabilityAck) GetAccepted() []string {
 }
 
 type CapabilityMessage struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	// Types that are assignable to Kind:
+	state protoimpl.MessageState `protogen:"open.v1"`
+	// Types that are valid to be assigned to Kind:
 	//
 	//	*CapabilityMessage_Offer
 	//	*CapabilityMessage_Request
 	//	*CapabilityMessage_Ack
-	Kind isCapabilityMessage_Kind `protobuf_oneof:"kind"`
+	Kind          isCapabilityMessage_Kind `protobuf_oneof:"kind"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CapabilityMessage) Reset() {
 	*x = CapabilityMessage{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[10]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[10]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CapabilityMessage) String() string {
@@ -976,7 +960,7 @@ func (*CapabilityMessage) ProtoMessage() {}
 
 func (x *CapabilityMessage) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[10]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -991,30 +975,36 @@ func (*CapabilityMessage) Descriptor() ([]byte, []int) {
 	return file_axcp_proto_rawDescGZIP(), []int{10}
 }
 
-func (m *CapabilityMessage) GetKind() isCapabilityMessage_Kind {
-	if m != nil {
-		return m.Kind
+func (x *CapabilityMessage) GetKind() isCapabilityMessage_Kind {
+	if x != nil {
+		return x.Kind
 	}
 	return nil
 }
 
 func (x *CapabilityMessage) GetOffer() *CapabilityOffer {
-	if x, ok := x.GetKind().(*CapabilityMessage_Offer); ok {
-		return x.Offer
+	if x != nil {
+		if x, ok := x.Kind.(*CapabilityMessage_Offer); ok {
+			return x.Offer
+		}
 	}
 	return nil
 }
 
 func (x *CapabilityMessage) GetRequest() *CapabilityRequest {
-	if x, ok := x.GetKind().(*CapabilityMessage_Request); ok {
-		return x.Request
+	if x != nil {
+		if x, ok := x.Kind.(*CapabilityMessage_Request); ok {
+			return x.Request
+		}
 	}
 	return nil
 }
 
 func (x *CapabilityMessage) GetAck() *CapabilityAck {
-	if x, ok := x.GetKind().(*CapabilityMessage_Ack); ok {
-		return x.Ack
+	if x != nil {
+		if x, ok := x.Kind.(*CapabilityMessage_Ack); ok {
+			return x.Ack
+		}
 	}
 	return nil
 }
@@ -1042,27 +1032,24 @@ func (*CapabilityMessage_Request) isCapabilityMessage_Kind() {}
 func (*CapabilityMessage_Ack) isCapabilityMessage_Kind() {}
 
 type CapabilityDescriptor struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	ToolId            string    `protobuf:"bytes,1,opt,name=tool_id,json=toolId,proto3" json:"tool_id,omitempty"`
-	InputSchema       string    `protobuf:"bytes,2,opt,name=input_schema,json=inputSchema,proto3" json:"input_schema,omitempty"`
-	OutputSchema      string    `protobuf:"bytes,3,opt,name=output_schema,json=outputSchema,proto3" json:"output_schema,omitempty"`
-	TimeoutMs         uint32    `protobuf:"varint,4,opt,name=timeout_ms,json=timeoutMs,proto3" json:"timeout_ms,omitempty"`
-	ResourceHint      string    `protobuf:"bytes,5,opt,name=resource_hint,json=resourceHint,proto3" json:"resource_hint,omitempty"`
-	AuthScope         []string  `protobuf:"bytes,6,rep,name=auth_scope,json=authScope,proto3" json:"auth_scope,omitempty"`
-	DescriptorVersion string    `protobuf:"bytes,7,opt,name=descriptor_version,json=descriptorVersion,proto3" json:"descriptor_version,omitempty"`
-	Dp                *DpParams `protobuf:"bytes,8,opt,name=dp,proto3" json:"dp,omitempty"` // profile ≥3
+	state             protoimpl.MessageState `protogen:"open.v1"`
+	ToolId            string                 `protobuf:"bytes,1,opt,name=tool_id,json=toolId,proto3" json:"tool_id,omitempty"`
+	InputSchema       string                 `protobuf:"bytes,2,opt,name=input_schema,json=inputSchema,proto3" json:"input_schema,omitempty"`
+	OutputSchema      string                 `protobuf:"bytes,3,opt,name=output_schema,json=outputSchema,proto3" json:"output_schema,omitempty"`
+	TimeoutMs         uint32                 `protobuf:"varint,4,opt,name=timeout_ms,json=timeoutMs,proto3" json:"timeout_ms,omitempty"`
+	ResourceHint      string                 `protobuf:"bytes,5,opt,name=resource_hint,json=resourceHint,proto3" json:"resource_hint,omitempty"`
+	AuthScope         []string               `protobuf:"bytes,6,rep,name=auth_scope,json=authScope,proto3" json:"auth_scope,omitempty"`
+	DescriptorVersion string                 `protobuf:"bytes,7,opt,name=descriptor_version,json=descriptorVersion,proto3" json:"descriptor_version,omitempty"`
+	Dp                *DpParams              `protobuf:"bytes,8,opt,name=dp,proto3" json:"dp,omitempty"` // profile ≥3
+	unknownFields     protoimpl.UnknownFields
+	sizeCache         protoimpl.SizeCache
 }
 
 func (x *CapabilityDescriptor) Reset() {
 	*x = CapabilityDescriptor{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[11]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[11]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CapabilityDescriptor) String() string {
@@ -1073,7 +1060,7 @@ func (*CapabilityDescriptor) ProtoMessage() {}
 
 func (x *CapabilityDescriptor) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[11]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1145,21 +1132,18 @@ func (x *CapabilityDescriptor) GetDp() *DpParams {
 }
 
 type ProfileNegotiate struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	SupportedMask uint32                 `protobuf:"varint,1,opt,name=supported_mask,json=supportedMask,proto3" json:"supported_mask,omitempty"` // bitmask; bit0=Profile-0 …
+	MinRequired   uint32                 `protobuf:"varint,2,opt,name=min_required,json=minRequired,proto3" json:"min_required,omitempty"`       // lowest acceptable profile
 	unknownFields protoimpl.UnknownFields
-
-	SupportedMask uint32 `protobuf:"varint,1,opt,name=supported_mask,json=supportedMask,proto3" json:"supported_mask,omitempty"` // bitmask; bit0=Profile-0 …
-	MinRequired   uint32 `protobuf:"varint,2,opt,name=min_required,json=minRequired,proto3" json:"min_required,omitempty"`       // lowest acceptable profile
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ProfileNegotiate) Reset() {
 	*x = ProfileNegotiate{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[12]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[12]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ProfileNegotiate) String() string {
@@ -1170,7 +1154,7 @@ func (*ProfileNegotiate) ProtoMessage() {}
 
 func (x *ProfileNegotiate) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[12]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1200,20 +1184,17 @@ func (x *ProfileNegotiate) GetMinRequired() uint32 {
 }
 
 type ProfileAck struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	AgreedProfile uint32                 `protobuf:"varint,1,opt,name=agreed_profile,json=agreedProfile,proto3" json:"agreed_profile,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	AgreedProfile uint32 `protobuf:"varint,1,opt,name=agreed_profile,json=agreedProfile,proto3" json:"agreed_profile,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ProfileAck) Reset() {
 	*x = ProfileAck{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[13]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[13]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ProfileAck) String() string {
@@ -1224,7 +1205,7 @@ func (*ProfileAck) ProtoMessage() {}
 
 func (x *ProfileAck) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[13]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1247,22 +1228,19 @@ func (x *ProfileAck) GetAgreedProfile() uint32 {
 }
 
 type RoutePolicyMessage struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	PolicyId      string                 `protobuf:"bytes,1,opt,name=policy_id,json=policyId,proto3" json:"policy_id,omitempty"`
+	WasmBlob      []byte                 `protobuf:"bytes,2,opt,name=wasm_blob,json=wasmBlob,proto3" json:"wasm_blob,omitempty"`
+	TtlMs         uint32                 `protobuf:"varint,3,opt,name=ttl_ms,json=ttlMs,proto3" json:"ttl_ms,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	PolicyId string `protobuf:"bytes,1,opt,name=policy_id,json=policyId,proto3" json:"policy_id,omitempty"`
-	WasmBlob []byte `protobuf:"bytes,2,opt,name=wasm_blob,json=wasmBlob,proto3" json:"wasm_blob,omitempty"`
-	TtlMs    uint32 `protobuf:"varint,3,opt,name=ttl_ms,json=ttlMs,proto3" json:"ttl_ms,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RoutePolicyMessage) Reset() {
 	*x = RoutePolicyMessage{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[14]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[14]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RoutePolicyMessage) String() string {
@@ -1273,7 +1251,7 @@ func (*RoutePolicyMessage) ProtoMessage() {}
 
 func (x *RoutePolicyMessage) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[14]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1310,22 +1288,19 @@ func (x *RoutePolicyMessage) GetTtlMs() uint32 {
 }
 
 type SystemStats struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	CpuPercent    uint32                 `protobuf:"varint,1,opt,name=cpu_percent,json=cpuPercent,proto3" json:"cpu_percent,omitempty"`
+	MemBytes      uint64                 `protobuf:"varint,2,opt,name=mem_bytes,json=memBytes,proto3" json:"mem_bytes,omitempty"`
+	TemperatureC  uint32                 `protobuf:"varint,3,opt,name=temperature_c,json=temperatureC,proto3" json:"temperature_c,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	CpuPercent   uint32 `protobuf:"varint,1,opt,name=cpu_percent,json=cpuPercent,proto3" json:"cpu_percent,omitempty"`
-	MemBytes     uint64 `protobuf:"varint,2,opt,name=mem_bytes,json=memBytes,proto3" json:"mem_bytes,omitempty"`
-	TemperatureC uint32 `protobuf:"varint,3,opt,name=temperature_c,json=temperatureC,proto3" json:"temperature_c,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *SystemStats) Reset() {
 	*x = SystemStats{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[15]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[15]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SystemStats) String() string {
@@ -1336,7 +1311,7 @@ func (*SystemStats) ProtoMessage() {}
 
 func (x *SystemStats) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[15]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1373,21 +1348,18 @@ func (x *SystemStats) GetTemperatureC() uint32 {
 }
 
 type TokenUsage struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	PromptTokens     uint32 `protobuf:"varint,1,opt,name=prompt_tokens,json=promptTokens,proto3" json:"prompt_tokens,omitempty"`
-	CompletionTokens uint32 `protobuf:"varint,2,opt,name=completion_tokens,json=completionTokens,proto3" json:"completion_tokens,omitempty"`
+	state            protoimpl.MessageState `protogen:"open.v1"`
+	PromptTokens     uint32                 `protobuf:"varint,1,opt,name=prompt_tokens,json=promptTokens,proto3" json:"prompt_tokens,omitempty"`
+	CompletionTokens uint32                 `protobuf:"varint,2,opt,name=completion_tokens,json=completionTokens,proto3" json:"completion_tokens,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *TokenUsage) Reset() {
 	*x = TokenUsage{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[16]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[16]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TokenUsage) String() string {
@@ -1398,7 +1370,7 @@ func (*TokenUsage) ProtoMessage() {}
 
 func (x *TokenUsage) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[16]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1428,25 +1400,22 @@ func (x *TokenUsage) GetCompletionTokens() uint32 {
 }
 
 type TelemetryDatagram struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	TimestampMs uint64 `protobuf:"varint,1,opt,name=timestamp_ms,json=timestampMs,proto3" json:"timestamp_ms,omitempty"`
-	// Types that are assignable to Payload:
+	state       protoimpl.MessageState `protogen:"open.v1"`
+	TimestampMs uint64                 `protobuf:"varint,1,opt,name=timestamp_ms,json=timestampMs,proto3" json:"timestamp_ms,omitempty"`
+	// Types that are valid to be assigned to Payload:
 	//
 	//	*TelemetryDatagram_System
 	//	*TelemetryDatagram_Tokens
-	Payload isTelemetryDatagram_Payload `protobuf_oneof:"payload"`
+	Payload       isTelemetryDatagram_Payload `protobuf_oneof:"payload"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TelemetryDatagram) Reset() {
 	*x = TelemetryDatagram{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[17]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[17]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TelemetryDatagram) String() string {
@@ -1457,7 +1426,7 @@ func (*TelemetryDatagram) ProtoMessage() {}
 
 func (x *TelemetryDatagram) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[17]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1479,23 +1448,27 @@ func (x *TelemetryDatagram) GetTimestampMs() uint64 {
 	return 0
 }
 
-func (m *TelemetryDatagram) GetPayload() isTelemetryDatagram_Payload {
-	if m != nil {
-		return m.Payload
+func (x *TelemetryDatagram) GetPayload() isTelemetryDatagram_Payload {
+	if x != nil {
+		return x.Payload
 	}
 	return nil
 }
 
 func (x *TelemetryDatagram) GetSystem() *SystemStats {
-	if x, ok := x.GetPayload().(*TelemetryDatagram_System); ok {
-		return x.System
+	if x != nil {
+		if x, ok := x.Payload.(*TelemetryDatagram_System); ok {
+			return x.System
+		}
 	}
 	return nil
 }
 
 func (x *TelemetryDatagram) GetTokens() *TokenUsage {
-	if x, ok := x.GetPayload().(*TelemetryDatagram_Tokens); ok {
-		return x.Tokens
+	if x != nil {
+		if x, ok := x.Payload.(*TelemetryDatagram_Tokens); ok {
+			return x.Tokens
+		}
 	}
 	return nil
 }
@@ -1517,24 +1490,21 @@ func (*TelemetryDatagram_System) isTelemetryDatagram_Payload() {}
 func (*TelemetryDatagram_Tokens) isTelemetryDatagram_Payload() {}
 
 type DpParams struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Epsilon       float64                `protobuf:"fixed64,1,opt,name=epsilon,proto3" json:"epsilon,omitempty"`
+	Delta         float64                `protobuf:"fixed64,2,opt,name=delta,proto3" json:"delta,omitempty"`
+	Mech          DpMechanism            `protobuf:"varint,3,opt,name=mech,proto3,enum=axcp.v0_1.DpMechanism" json:"mech,omitempty"`
+	ClipNorm      float64                `protobuf:"fixed64,4,opt,name=clip_norm,json=clipNorm,proto3" json:"clip_norm,omitempty"`
+	Gran          DpParams_Granularity   `protobuf:"varint,5,opt,name=gran,proto3,enum=axcp.v0_1.DpParams_Granularity" json:"gran,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Epsilon  float64              `protobuf:"fixed64,1,opt,name=epsilon,proto3" json:"epsilon,omitempty"`
-	Delta    float64              `protobuf:"fixed64,2,opt,name=delta,proto3" json:"delta,omitempty"`
-	Mech     DpMechanism          `protobuf:"varint,3,opt,name=mech,proto3,enum=axcp.v0_1.DpMechanism" json:"mech,omitempty"`
-	ClipNorm float64              `protobuf:"fixed64,4,opt,name=clip_norm,json=clipNorm,proto3" json:"clip_norm,omitempty"`
-	Gran     DpParams_Granularity `protobuf:"varint,5,opt,name=gran,proto3,enum=axcp.v0_1.DpParams_Granularity" json:"gran,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *DpParams) Reset() {
 	*x = DpParams{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[18]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[18]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *DpParams) String() string {
@@ -1545,7 +1515,7 @@ func (*DpParams) ProtoMessage() {}
 
 func (x *DpParams) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[18]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1596,22 +1566,19 @@ func (x *DpParams) GetGran() DpParams_Granularity {
 }
 
 type ErrorMessage struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Code          uint32                 `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`
+	Reason        string                 `protobuf:"bytes,2,opt,name=reason,proto3" json:"reason,omitempty"`
+	Diagnostics   []byte                 `protobuf:"bytes,3,opt,name=diagnostics,proto3" json:"diagnostics,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Code        uint32 `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`
-	Reason      string `protobuf:"bytes,2,opt,name=reason,proto3" json:"reason,omitempty"`
-	Diagnostics []byte `protobuf:"bytes,3,opt,name=diagnostics,proto3" json:"diagnostics,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ErrorMessage) Reset() {
 	*x = ErrorMessage{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[19]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[19]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ErrorMessage) String() string {
@@ -1622,7 +1589,7 @@ func (*ErrorMessage) ProtoMessage() {}
 
 func (x *ErrorMessage) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[19]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1659,20 +1626,17 @@ func (x *ErrorMessage) GetDiagnostics() []byte {
 }
 
 type McpJsonBlob struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Json          []byte                 `protobuf:"bytes,1,opt,name=json,proto3" json:"json,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Json []byte `protobuf:"bytes,1,opt,name=json,proto3" json:"json,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *McpJsonBlob) Reset() {
 	*x = McpJsonBlob{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[20]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[20]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *McpJsonBlob) String() string {
@@ -1683,7 +1647,7 @@ func (*McpJsonBlob) ProtoMessage() {}
 
 func (x *McpJsonBlob) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[20]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1706,20 +1670,17 @@ func (x *McpJsonBlob) GetJson() []byte {
 }
 
 type A2AJsonBlob struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Json          []byte                 `protobuf:"bytes,1,opt,name=json,proto3" json:"json,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Json []byte `protobuf:"bytes,1,opt,name=json,proto3" json:"json,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *A2AJsonBlob) Reset() {
 	*x = A2AJsonBlob{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_axcp_proto_msgTypes[21]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_axcp_proto_msgTypes[21]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *A2AJsonBlob) String() string {
@@ -1730,7 +1691,7 @@ func (*A2AJsonBlob) ProtoMessage() {}
 
 func (x *A2AJsonBlob) ProtoReflect() protoreflect.Message {
 	mi := &file_axcp_proto_msgTypes[21]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1754,241 +1715,158 @@ func (x *A2AJsonBlob) GetJson() []byte {
 
 var File_axcp_proto protoreflect.FileDescriptor
 
-var file_axcp_proto_rawDesc = []byte{
-	0x0a, 0x0a, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x09, 0x61, 0x78,
-	0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x22, 0x9a, 0x05, 0x0a, 0x0c, 0x41, 0x78, 0x63, 0x70,
-	0x45, 0x6e, 0x76, 0x65, 0x6c, 0x6f, 0x70, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65, 0x72, 0x73,
-	0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69,
-	0x6f, 0x6e, 0x12, 0x19, 0x0a, 0x08, 0x74, 0x72, 0x61, 0x63, 0x65, 0x5f, 0x69, 0x64, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x74, 0x72, 0x61, 0x63, 0x65, 0x49, 0x64, 0x12, 0x18, 0x0a,
-	0x07, 0x70, 0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x07,
-	0x70, 0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x12, 0x3e, 0x0a, 0x0d, 0x63, 0x6f, 0x6e, 0x74, 0x65,
-	0x78, 0x74, 0x5f, 0x70, 0x61, 0x74, 0x63, 0x68, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x17,
-	0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x43, 0x6f, 0x6e, 0x74, 0x65,
-	0x78, 0x74, 0x50, 0x61, 0x74, 0x63, 0x68, 0x48, 0x00, 0x52, 0x0c, 0x63, 0x6f, 0x6e, 0x74, 0x65,
-	0x78, 0x74, 0x50, 0x61, 0x74, 0x63, 0x68, 0x12, 0x45, 0x0a, 0x0e, 0x63, 0x61, 0x70, 0x61, 0x62,
-	0x69, 0x6c, 0x69, 0x74, 0x79, 0x5f, 0x6d, 0x73, 0x67, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x1c, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x43, 0x61, 0x70, 0x61,
-	0x62, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x48, 0x00, 0x52,
-	0x0d, 0x63, 0x61, 0x70, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x4d, 0x73, 0x67, 0x12, 0x3c,
-	0x0a, 0x09, 0x72, 0x6f, 0x75, 0x74, 0x65, 0x5f, 0x6d, 0x73, 0x67, 0x18, 0x06, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x1d, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x52, 0x6f,
-	0x75, 0x74, 0x65, 0x50, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
-	0x48, 0x00, 0x52, 0x08, 0x72, 0x6f, 0x75, 0x74, 0x65, 0x4d, 0x73, 0x67, 0x12, 0x2f, 0x0a, 0x05,
-	0x65, 0x72, 0x72, 0x6f, 0x72, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x17, 0x2e, 0x61, 0x78,
-	0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x4d, 0x65, 0x73,
-	0x73, 0x61, 0x67, 0x65, 0x48, 0x00, 0x52, 0x05, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x12, 0x3e, 0x0a,
-	0x0b, 0x70, 0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x5f, 0x6e, 0x65, 0x67, 0x18, 0x08, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x50,
-	0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x4e, 0x65, 0x67, 0x6f, 0x74, 0x69, 0x61, 0x74, 0x65, 0x48,
-	0x00, 0x52, 0x0a, 0x70, 0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x4e, 0x65, 0x67, 0x12, 0x38, 0x0a,
-	0x0b, 0x70, 0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x5f, 0x61, 0x63, 0x6b, 0x18, 0x09, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x15, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x50,
-	0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x41, 0x63, 0x6b, 0x48, 0x00, 0x52, 0x0a, 0x70, 0x72, 0x6f,
-	0x66, 0x69, 0x6c, 0x65, 0x41, 0x63, 0x6b, 0x12, 0x37, 0x0a, 0x09, 0x72, 0x65, 0x74, 0x72, 0x79,
-	0x5f, 0x65, 0x6e, 0x76, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x18, 0x2e, 0x61, 0x78, 0x63,
-	0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x52, 0x65, 0x74, 0x72, 0x79, 0x45, 0x6e, 0x76, 0x65,
-	0x6c, 0x6f, 0x70, 0x65, 0x48, 0x00, 0x52, 0x08, 0x72, 0x65, 0x74, 0x72, 0x79, 0x45, 0x6e, 0x76,
-	0x12, 0x3c, 0x0a, 0x09, 0x74, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x18, 0x0b, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e,
-	0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x44, 0x61, 0x74, 0x61, 0x67, 0x72, 0x61,
-	0x6d, 0x48, 0x00, 0x52, 0x09, 0x74, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x12, 0x1c,
-	0x0a, 0x09, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72, 0x65, 0x18, 0x64, 0x20, 0x01, 0x28,
-	0x0c, 0x52, 0x09, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72, 0x65, 0x12, 0x2b, 0x0a, 0x11,
-	0x61, 0x74, 0x74, 0x65, 0x73, 0x74, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x5f, 0x70, 0x72, 0x6f, 0x6f,
-	0x66, 0x18, 0x65, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x10, 0x61, 0x74, 0x74, 0x65, 0x73, 0x74, 0x61,
-	0x74, 0x69, 0x6f, 0x6e, 0x50, 0x72, 0x6f, 0x6f, 0x66, 0x42, 0x09, 0x0a, 0x07, 0x70, 0x61, 0x79,
-	0x6c, 0x6f, 0x61, 0x64, 0x22, 0xa3, 0x01, 0x0a, 0x07, 0x44, 0x65, 0x6c, 0x74, 0x61, 0x4f, 0x70,
-	0x12, 0x29, 0x0a, 0x02, 0x6f, 0x70, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x19, 0x2e, 0x61,
-	0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x44, 0x65, 0x6c, 0x74, 0x61, 0x4f, 0x70,
-	0x2e, 0x4f, 0x70, 0x54, 0x79, 0x70, 0x65, 0x52, 0x02, 0x6f, 0x70, 0x12, 0x12, 0x0a, 0x04, 0x70,
-	0x61, 0x74, 0x68, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x70, 0x61, 0x74, 0x68, 0x12,
-	0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x64,
-	0x61, 0x74, 0x61, 0x12, 0x0e, 0x0a, 0x02, 0x74, 0x73, 0x18, 0x04, 0x20, 0x01, 0x28, 0x04, 0x52,
-	0x02, 0x74, 0x73, 0x22, 0x35, 0x0a, 0x06, 0x4f, 0x70, 0x54, 0x79, 0x70, 0x65, 0x12, 0x07, 0x0a,
-	0x03, 0x41, 0x44, 0x44, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x52, 0x45, 0x50, 0x4c, 0x41, 0x43,
-	0x45, 0x10, 0x01, 0x12, 0x0a, 0x0a, 0x06, 0x52, 0x45, 0x4d, 0x4f, 0x56, 0x45, 0x10, 0x02, 0x12,
-	0x09, 0x0a, 0x05, 0x4d, 0x45, 0x52, 0x47, 0x45, 0x10, 0x03, 0x22, 0x76, 0x0a, 0x0c, 0x43, 0x6f,
-	0x6e, 0x74, 0x65, 0x78, 0x74, 0x50, 0x61, 0x74, 0x63, 0x68, 0x12, 0x1d, 0x0a, 0x0a, 0x63, 0x6f,
-	0x6e, 0x74, 0x65, 0x78, 0x74, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09,
-	0x63, 0x6f, 0x6e, 0x74, 0x65, 0x78, 0x74, 0x49, 0x64, 0x12, 0x21, 0x0a, 0x0c, 0x62, 0x61, 0x73,
-	0x65, 0x5f, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52,
-	0x0b, 0x62, 0x61, 0x73, 0x65, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x24, 0x0a, 0x03,
-	0x6f, 0x70, 0x73, 0x18, 0x03, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x12, 0x2e, 0x61, 0x78, 0x63, 0x70,
-	0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x44, 0x65, 0x6c, 0x74, 0x61, 0x4f, 0x70, 0x52, 0x03, 0x6f,
-	0x70, 0x73, 0x22, 0x4e, 0x0a, 0x13, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x78, 0x74, 0x47, 0x72, 0x61,
-	0x70, 0x68, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x1d, 0x0a, 0x0a, 0x63, 0x6f, 0x6e,
-	0x74, 0x65, 0x78, 0x74, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x63,
-	0x6f, 0x6e, 0x74, 0x65, 0x78, 0x74, 0x49, 0x64, 0x12, 0x18, 0x0a, 0x07, 0x76, 0x65, 0x72, 0x73,
-	0x69, 0x6f, 0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69,
-	0x6f, 0x6e, 0x22, 0x43, 0x0a, 0x0d, 0x53, 0x79, 0x6e, 0x63, 0x53, 0x75, 0x62, 0x73, 0x63, 0x72,
-	0x69, 0x62, 0x65, 0x12, 0x32, 0x0a, 0x04, 0x66, 0x72, 0x6f, 0x6d, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x1e, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x43, 0x6f,
-	0x6e, 0x74, 0x65, 0x78, 0x74, 0x47, 0x72, 0x61, 0x70, 0x68, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f,
-	0x6e, 0x52, 0x04, 0x66, 0x72, 0x6f, 0x6d, 0x22, 0x6f, 0x0a, 0x0b, 0x53, 0x79, 0x6e, 0x63, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x41, 0x0a, 0x0c, 0x6d, 0x69, 0x73, 0x73, 0x69, 0x6e,
-	0x67, 0x5f, 0x66, 0x72, 0x6f, 0x6d, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1e, 0x2e, 0x61,
-	0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x78, 0x74,
-	0x47, 0x72, 0x61, 0x70, 0x68, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x52, 0x0b, 0x6d, 0x69,
-	0x73, 0x73, 0x69, 0x6e, 0x67, 0x46, 0x72, 0x6f, 0x6d, 0x12, 0x1d, 0x0a, 0x0a, 0x74, 0x6f, 0x5f,
-	0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x09, 0x74,
-	0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x22, 0x6a, 0x0a, 0x0d, 0x52, 0x65, 0x74, 0x72,
-	0x79, 0x45, 0x6e, 0x76, 0x65, 0x6c, 0x6f, 0x70, 0x65, 0x12, 0x42, 0x0a, 0x10, 0x62, 0x75, 0x66,
-	0x66, 0x65, 0x72, 0x65, 0x64, 0x5f, 0x70, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x18, 0x01, 0x20,
-	0x03, 0x28, 0x0b, 0x32, 0x17, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e,
-	0x43, 0x6f, 0x6e, 0x74, 0x65, 0x78, 0x74, 0x50, 0x61, 0x74, 0x63, 0x68, 0x52, 0x0f, 0x62, 0x75,
-	0x66, 0x66, 0x65, 0x72, 0x65, 0x64, 0x50, 0x61, 0x74, 0x63, 0x68, 0x65, 0x73, 0x12, 0x15, 0x0a,
-	0x06, 0x74, 0x74, 0x6c, 0x5f, 0x6d, 0x73, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x05, 0x74,
-	0x74, 0x6c, 0x4d, 0x73, 0x22, 0x46, 0x0a, 0x0f, 0x43, 0x61, 0x70, 0x61, 0x62, 0x69, 0x6c, 0x69,
-	0x74, 0x79, 0x4f, 0x66, 0x66, 0x65, 0x72, 0x12, 0x33, 0x0a, 0x04, 0x64, 0x65, 0x73, 0x63, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1f, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f,
-	0x31, 0x2e, 0x43, 0x61, 0x70, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x44, 0x65, 0x73, 0x63,
-	0x72, 0x69, 0x70, 0x74, 0x6f, 0x72, 0x52, 0x04, 0x64, 0x65, 0x73, 0x63, 0x22, 0x25, 0x0a, 0x11,
-	0x43, 0x61, 0x70, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x12, 0x10, 0x0a, 0x03, 0x69, 0x64, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52, 0x03,
-	0x69, 0x64, 0x73, 0x22, 0x2b, 0x0a, 0x0d, 0x43, 0x61, 0x70, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74,
-	0x79, 0x41, 0x63, 0x6b, 0x12, 0x1a, 0x0a, 0x08, 0x61, 0x63, 0x63, 0x65, 0x70, 0x74, 0x65, 0x64,
-	0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52, 0x08, 0x61, 0x63, 0x63, 0x65, 0x70, 0x74, 0x65, 0x64,
-	0x22, 0xb7, 0x01, 0x0a, 0x11, 0x43, 0x61, 0x70, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x4d,
-	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x32, 0x0a, 0x05, 0x6f, 0x66, 0x66, 0x65, 0x72, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f,
-	0x31, 0x2e, 0x43, 0x61, 0x70, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x4f, 0x66, 0x66, 0x65,
-	0x72, 0x48, 0x00, 0x52, 0x05, 0x6f, 0x66, 0x66, 0x65, 0x72, 0x12, 0x38, 0x0a, 0x07, 0x72, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x61, 0x78,
-	0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x43, 0x61, 0x70, 0x61, 0x62, 0x69, 0x6c, 0x69,
-	0x74, 0x79, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x48, 0x00, 0x52, 0x07, 0x72, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x12, 0x2c, 0x0a, 0x03, 0x61, 0x63, 0x6b, 0x18, 0x03, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x18, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x43, 0x61,
-	0x70, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x41, 0x63, 0x6b, 0x48, 0x00, 0x52, 0x03, 0x61,
-	0x63, 0x6b, 0x42, 0x06, 0x0a, 0x04, 0x6b, 0x69, 0x6e, 0x64, 0x22, 0xae, 0x02, 0x0a, 0x14, 0x43,
-	0x61, 0x70, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x44, 0x65, 0x73, 0x63, 0x72, 0x69, 0x70,
-	0x74, 0x6f, 0x72, 0x12, 0x17, 0x0a, 0x07, 0x74, 0x6f, 0x6f, 0x6c, 0x5f, 0x69, 0x64, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x74, 0x6f, 0x6f, 0x6c, 0x49, 0x64, 0x12, 0x21, 0x0a, 0x0c,
-	0x69, 0x6e, 0x70, 0x75, 0x74, 0x5f, 0x73, 0x63, 0x68, 0x65, 0x6d, 0x61, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x0b, 0x69, 0x6e, 0x70, 0x75, 0x74, 0x53, 0x63, 0x68, 0x65, 0x6d, 0x61, 0x12,
-	0x23, 0x0a, 0x0d, 0x6f, 0x75, 0x74, 0x70, 0x75, 0x74, 0x5f, 0x73, 0x63, 0x68, 0x65, 0x6d, 0x61,
-	0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0c, 0x6f, 0x75, 0x74, 0x70, 0x75, 0x74, 0x53, 0x63,
-	0x68, 0x65, 0x6d, 0x61, 0x12, 0x1d, 0x0a, 0x0a, 0x74, 0x69, 0x6d, 0x65, 0x6f, 0x75, 0x74, 0x5f,
-	0x6d, 0x73, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x09, 0x74, 0x69, 0x6d, 0x65, 0x6f, 0x75,
-	0x74, 0x4d, 0x73, 0x12, 0x23, 0x0a, 0x0d, 0x72, 0x65, 0x73, 0x6f, 0x75, 0x72, 0x63, 0x65, 0x5f,
-	0x68, 0x69, 0x6e, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0c, 0x72, 0x65, 0x73, 0x6f,
-	0x75, 0x72, 0x63, 0x65, 0x48, 0x69, 0x6e, 0x74, 0x12, 0x1d, 0x0a, 0x0a, 0x61, 0x75, 0x74, 0x68,
-	0x5f, 0x73, 0x63, 0x6f, 0x70, 0x65, 0x18, 0x06, 0x20, 0x03, 0x28, 0x09, 0x52, 0x09, 0x61, 0x75,
-	0x74, 0x68, 0x53, 0x63, 0x6f, 0x70, 0x65, 0x12, 0x2d, 0x0a, 0x12, 0x64, 0x65, 0x73, 0x63, 0x72,
-	0x69, 0x70, 0x74, 0x6f, 0x72, 0x5f, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x07, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x11, 0x64, 0x65, 0x73, 0x63, 0x72, 0x69, 0x70, 0x74, 0x6f, 0x72, 0x56,
-	0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x23, 0x0a, 0x02, 0x64, 0x70, 0x18, 0x08, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x13, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x44,
-	0x70, 0x50, 0x61, 0x72, 0x61, 0x6d, 0x73, 0x52, 0x02, 0x64, 0x70, 0x22, 0x5c, 0x0a, 0x10, 0x50,
-	0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x4e, 0x65, 0x67, 0x6f, 0x74, 0x69, 0x61, 0x74, 0x65, 0x12,
-	0x25, 0x0a, 0x0e, 0x73, 0x75, 0x70, 0x70, 0x6f, 0x72, 0x74, 0x65, 0x64, 0x5f, 0x6d, 0x61, 0x73,
-	0x6b, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0d, 0x73, 0x75, 0x70, 0x70, 0x6f, 0x72, 0x74,
-	0x65, 0x64, 0x4d, 0x61, 0x73, 0x6b, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x69, 0x6e, 0x5f, 0x72, 0x65,
-	0x71, 0x75, 0x69, 0x72, 0x65, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x6d, 0x69,
-	0x6e, 0x52, 0x65, 0x71, 0x75, 0x69, 0x72, 0x65, 0x64, 0x22, 0x33, 0x0a, 0x0a, 0x50, 0x72, 0x6f,
-	0x66, 0x69, 0x6c, 0x65, 0x41, 0x63, 0x6b, 0x12, 0x25, 0x0a, 0x0e, 0x61, 0x67, 0x72, 0x65, 0x65,
-	0x64, 0x5f, 0x70, 0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52,
-	0x0d, 0x61, 0x67, 0x72, 0x65, 0x65, 0x64, 0x50, 0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x22, 0x65,
-	0x0a, 0x12, 0x52, 0x6f, 0x75, 0x74, 0x65, 0x50, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x4d, 0x65, 0x73,
-	0x73, 0x61, 0x67, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x70, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x5f, 0x69,
-	0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x70, 0x6f, 0x6c, 0x69, 0x63, 0x79, 0x49,
-	0x64, 0x12, 0x1b, 0x0a, 0x09, 0x77, 0x61, 0x73, 0x6d, 0x5f, 0x62, 0x6c, 0x6f, 0x62, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x0c, 0x52, 0x08, 0x77, 0x61, 0x73, 0x6d, 0x42, 0x6c, 0x6f, 0x62, 0x12, 0x15,
-	0x0a, 0x06, 0x74, 0x74, 0x6c, 0x5f, 0x6d, 0x73, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x05,
-	0x74, 0x74, 0x6c, 0x4d, 0x73, 0x22, 0x70, 0x0a, 0x0b, 0x53, 0x79, 0x73, 0x74, 0x65, 0x6d, 0x53,
-	0x74, 0x61, 0x74, 0x73, 0x12, 0x1f, 0x0a, 0x0b, 0x63, 0x70, 0x75, 0x5f, 0x70, 0x65, 0x72, 0x63,
-	0x65, 0x6e, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0a, 0x63, 0x70, 0x75, 0x50, 0x65,
-	0x72, 0x63, 0x65, 0x6e, 0x74, 0x12, 0x1b, 0x0a, 0x09, 0x6d, 0x65, 0x6d, 0x5f, 0x62, 0x79, 0x74,
-	0x65, 0x73, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x08, 0x6d, 0x65, 0x6d, 0x42, 0x79, 0x74,
-	0x65, 0x73, 0x12, 0x23, 0x0a, 0x0d, 0x74, 0x65, 0x6d, 0x70, 0x65, 0x72, 0x61, 0x74, 0x75, 0x72,
-	0x65, 0x5f, 0x63, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0c, 0x74, 0x65, 0x6d, 0x70, 0x65,
-	0x72, 0x61, 0x74, 0x75, 0x72, 0x65, 0x43, 0x22, 0x5e, 0x0a, 0x0a, 0x54, 0x6f, 0x6b, 0x65, 0x6e,
-	0x55, 0x73, 0x61, 0x67, 0x65, 0x12, 0x23, 0x0a, 0x0d, 0x70, 0x72, 0x6f, 0x6d, 0x70, 0x74, 0x5f,
-	0x74, 0x6f, 0x6b, 0x65, 0x6e, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0c, 0x70, 0x72,
-	0x6f, 0x6d, 0x70, 0x74, 0x54, 0x6f, 0x6b, 0x65, 0x6e, 0x73, 0x12, 0x2b, 0x0a, 0x11, 0x63, 0x6f,
-	0x6d, 0x70, 0x6c, 0x65, 0x74, 0x69, 0x6f, 0x6e, 0x5f, 0x74, 0x6f, 0x6b, 0x65, 0x6e, 0x73, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x10, 0x63, 0x6f, 0x6d, 0x70, 0x6c, 0x65, 0x74, 0x69, 0x6f,
-	0x6e, 0x54, 0x6f, 0x6b, 0x65, 0x6e, 0x73, 0x22, 0xa4, 0x01, 0x0a, 0x11, 0x54, 0x65, 0x6c, 0x65,
-	0x6d, 0x65, 0x74, 0x72, 0x79, 0x44, 0x61, 0x74, 0x61, 0x67, 0x72, 0x61, 0x6d, 0x12, 0x21, 0x0a,
-	0x0c, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x5f, 0x6d, 0x73, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x04, 0x52, 0x0b, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x4d, 0x73,
-	0x12, 0x30, 0x0a, 0x06, 0x73, 0x79, 0x73, 0x74, 0x65, 0x6d, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x16, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x53, 0x79, 0x73,
-	0x74, 0x65, 0x6d, 0x53, 0x74, 0x61, 0x74, 0x73, 0x48, 0x00, 0x52, 0x06, 0x73, 0x79, 0x73, 0x74,
-	0x65, 0x6d, 0x12, 0x2f, 0x0a, 0x06, 0x74, 0x6f, 0x6b, 0x65, 0x6e, 0x73, 0x18, 0x0b, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x15, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x54,
-	0x6f, 0x6b, 0x65, 0x6e, 0x55, 0x73, 0x61, 0x67, 0x65, 0x48, 0x00, 0x52, 0x06, 0x74, 0x6f, 0x6b,
-	0x65, 0x6e, 0x73, 0x42, 0x09, 0x0a, 0x07, 0x70, 0x61, 0x79, 0x6c, 0x6f, 0x61, 0x64, 0x22, 0xe9,
-	0x01, 0x0a, 0x08, 0x44, 0x70, 0x50, 0x61, 0x72, 0x61, 0x6d, 0x73, 0x12, 0x18, 0x0a, 0x07, 0x65,
-	0x70, 0x73, 0x69, 0x6c, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x01, 0x52, 0x07, 0x65, 0x70,
-	0x73, 0x69, 0x6c, 0x6f, 0x6e, 0x12, 0x14, 0x0a, 0x05, 0x64, 0x65, 0x6c, 0x74, 0x61, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x01, 0x52, 0x05, 0x64, 0x65, 0x6c, 0x74, 0x61, 0x12, 0x2a, 0x0a, 0x04, 0x6d,
-	0x65, 0x63, 0x68, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x16, 0x2e, 0x61, 0x78, 0x63, 0x70,
-	0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x44, 0x70, 0x4d, 0x65, 0x63, 0x68, 0x61, 0x6e, 0x69, 0x73,
-	0x6d, 0x52, 0x04, 0x6d, 0x65, 0x63, 0x68, 0x12, 0x1b, 0x0a, 0x09, 0x63, 0x6c, 0x69, 0x70, 0x5f,
-	0x6e, 0x6f, 0x72, 0x6d, 0x18, 0x04, 0x20, 0x01, 0x28, 0x01, 0x52, 0x08, 0x63, 0x6c, 0x69, 0x70,
-	0x4e, 0x6f, 0x72, 0x6d, 0x12, 0x33, 0x0a, 0x04, 0x67, 0x72, 0x61, 0x6e, 0x18, 0x05, 0x20, 0x01,
-	0x28, 0x0e, 0x32, 0x1f, 0x2e, 0x61, 0x78, 0x63, 0x70, 0x2e, 0x76, 0x30, 0x5f, 0x31, 0x2e, 0x44,
-	0x70, 0x50, 0x61, 0x72, 0x61, 0x6d, 0x73, 0x2e, 0x47, 0x72, 0x61, 0x6e, 0x75, 0x6c, 0x61, 0x72,
-	0x69, 0x74, 0x79, 0x52, 0x04, 0x67, 0x72, 0x61, 0x6e, 0x22, 0x2f, 0x0a, 0x0b, 0x47, 0x72, 0x61,
-	0x6e, 0x75, 0x6c, 0x61, 0x72, 0x69, 0x74, 0x79, 0x12, 0x0a, 0x0a, 0x06, 0x52, 0x45, 0x43, 0x4f,
-	0x52, 0x44, 0x10, 0x00, 0x12, 0x09, 0x0a, 0x05, 0x42, 0x41, 0x54, 0x43, 0x48, 0x10, 0x01, 0x12,
-	0x09, 0x0a, 0x05, 0x54, 0x4f, 0x4b, 0x45, 0x4e, 0x10, 0x02, 0x22, 0x5c, 0x0a, 0x0c, 0x45, 0x72,
-	0x72, 0x6f, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x6f,
-	0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x12, 0x16,
-	0x0a, 0x06, 0x72, 0x65, 0x61, 0x73, 0x6f, 0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06,
-	0x72, 0x65, 0x61, 0x73, 0x6f, 0x6e, 0x12, 0x20, 0x0a, 0x0b, 0x64, 0x69, 0x61, 0x67, 0x6e, 0x6f,
-	0x73, 0x74, 0x69, 0x63, 0x73, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0b, 0x64, 0x69, 0x61,
-	0x67, 0x6e, 0x6f, 0x73, 0x74, 0x69, 0x63, 0x73, 0x22, 0x21, 0x0a, 0x0b, 0x4d, 0x63, 0x70, 0x4a,
-	0x73, 0x6f, 0x6e, 0x42, 0x6c, 0x6f, 0x62, 0x12, 0x12, 0x0a, 0x04, 0x6a, 0x73, 0x6f, 0x6e, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x6a, 0x73, 0x6f, 0x6e, 0x22, 0x21, 0x0a, 0x0b, 0x41,
-	0x32, 0x41, 0x4a, 0x73, 0x6f, 0x6e, 0x42, 0x6c, 0x6f, 0x62, 0x12, 0x12, 0x0a, 0x04, 0x6a, 0x73,
-	0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x6a, 0x73, 0x6f, 0x6e, 0x2a, 0x28,
-	0x0a, 0x0b, 0x44, 0x70, 0x4d, 0x65, 0x63, 0x68, 0x61, 0x6e, 0x69, 0x73, 0x6d, 0x12, 0x0b, 0x0a,
-	0x07, 0x4c, 0x41, 0x50, 0x4c, 0x41, 0x43, 0x45, 0x10, 0x00, 0x12, 0x0c, 0x0a, 0x08, 0x47, 0x41,
-	0x55, 0x53, 0x53, 0x49, 0x41, 0x4e, 0x10, 0x01, 0x2a, 0xcd, 0x02, 0x0a, 0x09, 0x45, 0x72, 0x72,
-	0x6f, 0x72, 0x43, 0x6f, 0x64, 0x65, 0x12, 0x0b, 0x0a, 0x07, 0x55, 0x4e, 0x4b, 0x4e, 0x4f, 0x57,
-	0x4e, 0x10, 0x00, 0x12, 0x13, 0x0a, 0x0f, 0x49, 0x4e, 0x56, 0x41, 0x4c, 0x49, 0x44, 0x5f, 0x43,
-	0x4f, 0x4e, 0x54, 0x45, 0x58, 0x54, 0x10, 0x01, 0x12, 0x10, 0x0a, 0x0c, 0x55, 0x4e, 0x41, 0x55,
-	0x54, 0x48, 0x4f, 0x52, 0x49, 0x5a, 0x45, 0x44, 0x10, 0x02, 0x12, 0x12, 0x0a, 0x0e, 0x54, 0x4f,
-	0x4f, 0x4c, 0x5f, 0x4e, 0x4f, 0x54, 0x5f, 0x46, 0x4f, 0x55, 0x4e, 0x44, 0x10, 0x03, 0x12, 0x0b,
-	0x0a, 0x07, 0x54, 0x49, 0x4d, 0x45, 0x4f, 0x55, 0x54, 0x10, 0x04, 0x12, 0x17, 0x0a, 0x13, 0x55,
-	0x4e, 0x53, 0x55, 0x50, 0x50, 0x4f, 0x52, 0x54, 0x45, 0x44, 0x5f, 0x56, 0x45, 0x52, 0x53, 0x49,
-	0x4f, 0x4e, 0x10, 0x05, 0x12, 0x0d, 0x0a, 0x09, 0x42, 0x41, 0x44, 0x5f, 0x44, 0x45, 0x4c, 0x54,
-	0x41, 0x10, 0x06, 0x12, 0x15, 0x0a, 0x11, 0x50, 0x41, 0x59, 0x4c, 0x4f, 0x41, 0x44, 0x5f, 0x54,
-	0x4f, 0x4f, 0x5f, 0x4c, 0x41, 0x52, 0x47, 0x45, 0x10, 0x07, 0x12, 0x15, 0x0a, 0x11, 0x4d, 0x41,
-	0x4c, 0x46, 0x4f, 0x52, 0x4d, 0x45, 0x44, 0x5f, 0x52, 0x45, 0x51, 0x55, 0x45, 0x53, 0x54, 0x10,
-	0x08, 0x12, 0x15, 0x0a, 0x11, 0x54, 0x4f, 0x4f, 0x5f, 0x4d, 0x41, 0x4e, 0x59, 0x5f, 0x52, 0x45,
-	0x51, 0x55, 0x45, 0x53, 0x54, 0x53, 0x10, 0x09, 0x12, 0x14, 0x0a, 0x10, 0x50, 0x52, 0x4f, 0x46,
-	0x49, 0x4c, 0x45, 0x5f, 0x4d, 0x49, 0x53, 0x4d, 0x41, 0x54, 0x43, 0x48, 0x10, 0x0c, 0x12, 0x17,
-	0x0a, 0x13, 0x50, 0x52, 0x4f, 0x46, 0x49, 0x4c, 0x45, 0x5f, 0x55, 0x4e, 0x53, 0x55, 0x50, 0x50,
-	0x4f, 0x52, 0x54, 0x45, 0x44, 0x10, 0x0d, 0x12, 0x1e, 0x0a, 0x1a, 0x50, 0x52, 0x4f, 0x46, 0x49,
-	0x4c, 0x45, 0x5f, 0x4e, 0x45, 0x47, 0x4f, 0x54, 0x49, 0x41, 0x54, 0x49, 0x4f, 0x4e, 0x5f, 0x46,
-	0x41, 0x49, 0x4c, 0x45, 0x44, 0x10, 0x0e, 0x12, 0x17, 0x0a, 0x13, 0x4d, 0x49, 0x53, 0x53, 0x49,
-	0x4e, 0x47, 0x5f, 0x50, 0x41, 0x54, 0x43, 0x48, 0x5f, 0x52, 0x41, 0x4e, 0x47, 0x45, 0x10, 0x0f,
-	0x12, 0x16, 0x0a, 0x12, 0x44, 0x50, 0x5f, 0x50, 0x4f, 0x4c, 0x49, 0x43, 0x59, 0x5f, 0x43, 0x4f,
-	0x4e, 0x46, 0x4c, 0x49, 0x43, 0x54, 0x10, 0x10, 0x42, 0x36, 0x5a, 0x34, 0x67, 0x69, 0x74, 0x68,
-	0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x74, 0x72, 0x61, 0x64, 0x65, 0x70, 0x68, 0x61, 0x6e,
-	0x74, 0x6f, 0x6d, 0x2f, 0x61, 0x78, 0x63, 0x70, 0x2d, 0x73, 0x70, 0x65, 0x63, 0x2f, 0x73, 0x64,
-	0x6b, 0x2f, 0x67, 0x6f, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x70, 0x62,
-	0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_axcp_proto_rawDesc = "" +
+	"\n" +
+	"\n" +
+	"axcp.proto\x12\taxcp.v0_1\"\x9a\x05\n" +
+	"\fAxcpEnvelope\x12\x18\n" +
+	"\aversion\x18\x01 \x01(\rR\aversion\x12\x19\n" +
+	"\btrace_id\x18\x02 \x01(\tR\atraceId\x12\x18\n" +
+	"\aprofile\x18\x03 \x01(\rR\aprofile\x12>\n" +
+	"\rcontext_patch\x18\x04 \x01(\v2\x17.axcp.v0_1.ContextPatchH\x00R\fcontextPatch\x12E\n" +
+	"\x0ecapability_msg\x18\x05 \x01(\v2\x1c.axcp.v0_1.CapabilityMessageH\x00R\rcapabilityMsg\x12<\n" +
+	"\troute_msg\x18\x06 \x01(\v2\x1d.axcp.v0_1.RoutePolicyMessageH\x00R\brouteMsg\x12/\n" +
+	"\x05error\x18\a \x01(\v2\x17.axcp.v0_1.ErrorMessageH\x00R\x05error\x12>\n" +
+	"\vprofile_neg\x18\b \x01(\v2\x1b.axcp.v0_1.ProfileNegotiateH\x00R\n" +
+	"profileNeg\x128\n" +
+	"\vprofile_ack\x18\t \x01(\v2\x15.axcp.v0_1.ProfileAckH\x00R\n" +
+	"profileAck\x127\n" +
+	"\tretry_env\x18\n" +
+	" \x01(\v2\x18.axcp.v0_1.RetryEnvelopeH\x00R\bretryEnv\x12<\n" +
+	"\ttelemetry\x18\v \x01(\v2\x1c.axcp.v0_1.TelemetryDatagramH\x00R\ttelemetry\x12\x1c\n" +
+	"\tsignature\x18d \x01(\fR\tsignature\x12+\n" +
+	"\x11attestation_proof\x18e \x01(\fR\x10attestationProofB\t\n" +
+	"\apayload\"\xa3\x01\n" +
+	"\aDeltaOp\x12)\n" +
+	"\x02op\x18\x01 \x01(\x0e2\x19.axcp.v0_1.DeltaOp.OpTypeR\x02op\x12\x12\n" +
+	"\x04path\x18\x02 \x01(\tR\x04path\x12\x12\n" +
+	"\x04data\x18\x03 \x01(\fR\x04data\x12\x0e\n" +
+	"\x02ts\x18\x04 \x01(\x04R\x02ts\"5\n" +
+	"\x06OpType\x12\a\n" +
+	"\x03ADD\x10\x00\x12\v\n" +
+	"\aREPLACE\x10\x01\x12\n" +
+	"\n" +
+	"\x06REMOVE\x10\x02\x12\t\n" +
+	"\x05MERGE\x10\x03\"v\n" +
+	"\fContextPatch\x12\x1d\n" +
+	"\n" +
+	"context_id\x18\x01 \x01(\tR\tcontextId\x12!\n" +
+	"\fbase_version\x18\x02 \x01(\x04R\vbaseVersion\x12$\n" +
+	"\x03ops\x18\x03 \x03(\v2\x12.axcp.v0_1.DeltaOpR\x03ops\"N\n" +
+	"\x13ContextGraphVersion\x12\x1d\n" +
+	"\n" +
+	"context_id\x18\x01 \x01(\tR\tcontextId\x12\x18\n" +
+	"\aversion\x18\x02 \x01(\x04R\aversion\"C\n" +
+	"\rSyncSubscribe\x122\n" +
+	"\x04from\x18\x01 \x01(\v2\x1e.axcp.v0_1.ContextGraphVersionR\x04from\"o\n" +
+	"\vSyncRequest\x12A\n" +
+	"\fmissing_from\x18\x01 \x01(\v2\x1e.axcp.v0_1.ContextGraphVersionR\vmissingFrom\x12\x1d\n" +
+	"\n" +
+	"to_version\x18\x02 \x01(\x04R\ttoVersion\"j\n" +
+	"\rRetryEnvelope\x12B\n" +
+	"\x10buffered_patches\x18\x01 \x03(\v2\x17.axcp.v0_1.ContextPatchR\x0fbufferedPatches\x12\x15\n" +
+	"\x06ttl_ms\x18\x02 \x01(\rR\x05ttlMs\"F\n" +
+	"\x0fCapabilityOffer\x123\n" +
+	"\x04desc\x18\x01 \x01(\v2\x1f.axcp.v0_1.CapabilityDescriptorR\x04desc\"%\n" +
+	"\x11CapabilityRequest\x12\x10\n" +
+	"\x03ids\x18\x01 \x03(\tR\x03ids\"+\n" +
+	"\rCapabilityAck\x12\x1a\n" +
+	"\baccepted\x18\x01 \x03(\tR\baccepted\"\xb7\x01\n" +
+	"\x11CapabilityMessage\x122\n" +
+	"\x05offer\x18\x01 \x01(\v2\x1a.axcp.v0_1.CapabilityOfferH\x00R\x05offer\x128\n" +
+	"\arequest\x18\x02 \x01(\v2\x1c.axcp.v0_1.CapabilityRequestH\x00R\arequest\x12,\n" +
+	"\x03ack\x18\x03 \x01(\v2\x18.axcp.v0_1.CapabilityAckH\x00R\x03ackB\x06\n" +
+	"\x04kind\"\xae\x02\n" +
+	"\x14CapabilityDescriptor\x12\x17\n" +
+	"\atool_id\x18\x01 \x01(\tR\x06toolId\x12!\n" +
+	"\finput_schema\x18\x02 \x01(\tR\vinputSchema\x12#\n" +
+	"\routput_schema\x18\x03 \x01(\tR\foutputSchema\x12\x1d\n" +
+	"\n" +
+	"timeout_ms\x18\x04 \x01(\rR\ttimeoutMs\x12#\n" +
+	"\rresource_hint\x18\x05 \x01(\tR\fresourceHint\x12\x1d\n" +
+	"\n" +
+	"auth_scope\x18\x06 \x03(\tR\tauthScope\x12-\n" +
+	"\x12descriptor_version\x18\a \x01(\tR\x11descriptorVersion\x12#\n" +
+	"\x02dp\x18\b \x01(\v2\x13.axcp.v0_1.DpParamsR\x02dp\"\\\n" +
+	"\x10ProfileNegotiate\x12%\n" +
+	"\x0esupported_mask\x18\x01 \x01(\rR\rsupportedMask\x12!\n" +
+	"\fmin_required\x18\x02 \x01(\rR\vminRequired\"3\n" +
+	"\n" +
+	"ProfileAck\x12%\n" +
+	"\x0eagreed_profile\x18\x01 \x01(\rR\ragreedProfile\"e\n" +
+	"\x12RoutePolicyMessage\x12\x1b\n" +
+	"\tpolicy_id\x18\x01 \x01(\tR\bpolicyId\x12\x1b\n" +
+	"\twasm_blob\x18\x02 \x01(\fR\bwasmBlob\x12\x15\n" +
+	"\x06ttl_ms\x18\x03 \x01(\rR\x05ttlMs\"p\n" +
+	"\vSystemStats\x12\x1f\n" +
+	"\vcpu_percent\x18\x01 \x01(\rR\n" +
+	"cpuPercent\x12\x1b\n" +
+	"\tmem_bytes\x18\x02 \x01(\x04R\bmemBytes\x12#\n" +
+	"\rtemperature_c\x18\x03 \x01(\rR\ftemperatureC\"^\n" +
+	"\n" +
+	"TokenUsage\x12#\n" +
+	"\rprompt_tokens\x18\x01 \x01(\rR\fpromptTokens\x12+\n" +
+	"\x11completion_tokens\x18\x02 \x01(\rR\x10completionTokens\"\xa4\x01\n" +
+	"\x11TelemetryDatagram\x12!\n" +
+	"\ftimestamp_ms\x18\x01 \x01(\x04R\vtimestampMs\x120\n" +
+	"\x06system\x18\n" +
+	" \x01(\v2\x16.axcp.v0_1.SystemStatsH\x00R\x06system\x12/\n" +
+	"\x06tokens\x18\v \x01(\v2\x15.axcp.v0_1.TokenUsageH\x00R\x06tokensB\t\n" +
+	"\apayload\"\xe9\x01\n" +
+	"\bDpParams\x12\x18\n" +
+	"\aepsilon\x18\x01 \x01(\x01R\aepsilon\x12\x14\n" +
+	"\x05delta\x18\x02 \x01(\x01R\x05delta\x12*\n" +
+	"\x04mech\x18\x03 \x01(\x0e2\x16.axcp.v0_1.DpMechanismR\x04mech\x12\x1b\n" +
+	"\tclip_norm\x18\x04 \x01(\x01R\bclipNorm\x123\n" +
+	"\x04gran\x18\x05 \x01(\x0e2\x1f.axcp.v0_1.DpParams.GranularityR\x04gran\"/\n" +
+	"\vGranularity\x12\n" +
+	"\n" +
+	"\x06RECORD\x10\x00\x12\t\n" +
+	"\x05BATCH\x10\x01\x12\t\n" +
+	"\x05TOKEN\x10\x02\"\\\n" +
+	"\fErrorMessage\x12\x12\n" +
+	"\x04code\x18\x01 \x01(\rR\x04code\x12\x16\n" +
+	"\x06reason\x18\x02 \x01(\tR\x06reason\x12 \n" +
+	"\vdiagnostics\x18\x03 \x01(\fR\vdiagnostics\"!\n" +
+	"\vMcpJsonBlob\x12\x12\n" +
+	"\x04json\x18\x01 \x01(\fR\x04json\"!\n" +
+	"\vA2AJsonBlob\x12\x12\n" +
+	"\x04json\x18\x01 \x01(\fR\x04json*(\n" +
+	"\vDpMechanism\x12\v\n" +
+	"\aLAPLACE\x10\x00\x12\f\n" +
+	"\bGAUSSIAN\x10\x01*\xcd\x02\n" +
+	"\tErrorCode\x12\v\n" +
+	"\aUNKNOWN\x10\x00\x12\x13\n" +
+	"\x0fINVALID_CONTEXT\x10\x01\x12\x10\n" +
+	"\fUNAUTHORIZED\x10\x02\x12\x12\n" +
+	"\x0eTOOL_NOT_FOUND\x10\x03\x12\v\n" +
+	"\aTIMEOUT\x10\x04\x12\x17\n" +
+	"\x13UNSUPPORTED_VERSION\x10\x05\x12\r\n" +
+	"\tBAD_DELTA\x10\x06\x12\x15\n" +
+	"\x11PAYLOAD_TOO_LARGE\x10\a\x12\x15\n" +
+	"\x11MALFORMED_REQUEST\x10\b\x12\x15\n" +
+	"\x11TOO_MANY_REQUESTS\x10\t\x12\x14\n" +
+	"\x10PROFILE_MISMATCH\x10\f\x12\x17\n" +
+	"\x13PROFILE_UNSUPPORTED\x10\r\x12\x1e\n" +
+	"\x1aPROFILE_NEGOTIATION_FAILED\x10\x0e\x12\x17\n" +
+	"\x13MISSING_PATCH_RANGE\x10\x0f\x12\x16\n" +
+	"\x12DP_POLICY_CONFLICT\x10\x10B6Z4github.com/tradephantom/axcp-spec/sdk/go/internal/pbb\x06proto3"
 
 var (
 	file_axcp_proto_rawDescOnce sync.Once
-	file_axcp_proto_rawDescData = file_axcp_proto_rawDesc
+	file_axcp_proto_rawDescData []byte
 )
 
 func file_axcp_proto_rawDescGZIP() []byte {
 	file_axcp_proto_rawDescOnce.Do(func() {
-		file_axcp_proto_rawDescData = protoimpl.X.CompressGZIP(file_axcp_proto_rawDescData)
+		file_axcp_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_axcp_proto_rawDesc), len(file_axcp_proto_rawDesc)))
 	})
 	return file_axcp_proto_rawDescData
 }
 
 var file_axcp_proto_enumTypes = make([]protoimpl.EnumInfo, 4)
 var file_axcp_proto_msgTypes = make([]protoimpl.MessageInfo, 22)
-var file_axcp_proto_goTypes = []interface{}{
+var file_axcp_proto_goTypes = []any{
 	(DpMechanism)(0),             // 0: axcp.v0_1.DpMechanism
 	(ErrorCode)(0),               // 1: axcp.v0_1.ErrorCode
 	(DeltaOp_OpType)(0),          // 2: axcp.v0_1.DeltaOp.OpType
@@ -2051,273 +1929,7 @@ func file_axcp_proto_init() {
 	if File_axcp_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_axcp_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*AxcpEnvelope); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*DeltaOp); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ContextPatch); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ContextGraphVersion); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SyncSubscribe); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SyncRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RetryEnvelope); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CapabilityOffer); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CapabilityRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CapabilityAck); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CapabilityMessage); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CapabilityDescriptor); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ProfileNegotiate); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ProfileAck); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RoutePolicyMessage); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SystemStats); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TokenUsage); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TelemetryDatagram); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[18].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*DpParams); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[19].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ErrorMessage); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[20].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*McpJsonBlob); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_axcp_proto_msgTypes[21].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*A2AJsonBlob); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_axcp_proto_msgTypes[0].OneofWrappers = []interface{}{
+	file_axcp_proto_msgTypes[0].OneofWrappers = []any{
 		(*AxcpEnvelope_ContextPatch)(nil),
 		(*AxcpEnvelope_CapabilityMsg)(nil),
 		(*AxcpEnvelope_RouteMsg)(nil),
@@ -2327,12 +1939,12 @@ func file_axcp_proto_init() {
 		(*AxcpEnvelope_RetryEnv)(nil),
 		(*AxcpEnvelope_Telemetry)(nil),
 	}
-	file_axcp_proto_msgTypes[10].OneofWrappers = []interface{}{
+	file_axcp_proto_msgTypes[10].OneofWrappers = []any{
 		(*CapabilityMessage_Offer)(nil),
 		(*CapabilityMessage_Request)(nil),
 		(*CapabilityMessage_Ack)(nil),
 	}
-	file_axcp_proto_msgTypes[17].OneofWrappers = []interface{}{
+	file_axcp_proto_msgTypes[17].OneofWrappers = []any{
 		(*TelemetryDatagram_System)(nil),
 		(*TelemetryDatagram_Tokens)(nil),
 	}
@@ -2340,7 +1952,7 @@ func file_axcp_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_axcp_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_axcp_proto_rawDesc), len(file_axcp_proto_rawDesc)),
 			NumEnums:      4,
 			NumMessages:   22,
 			NumExtensions: 0,
@@ -2352,7 +1964,6 @@ func file_axcp_proto_init() {
 		MessageInfos:      file_axcp_proto_msgTypes,
 	}.Build()
 	File_axcp_proto = out.File
-	file_axcp_proto_rawDesc = nil
 	file_axcp_proto_goTypes = nil
 	file_axcp_proto_depIdxs = nil
 }
